Splitting `Locator` and `PhpParserLocator`, which are now an interface and an implementation respectively
